Work on authorization rules and user login

+++ app/controllers/items_controller.rb
@@ -5,32 +5,37 @@ class ItemsController < ApplicationController
     else
       @items = Item.where("type = ?", params[:type]).first(12)
     end
+    authorize Item.new
   end
   
   def show
     @item = get_by_uid params[:uid]
+    authorize @item.becomes(Item)
   end
   
   # Redirects to /items/[:uid]#edit,
   # editing form is then handled by javascript.
   def edit
     @item = get_by_uid params[:uid]
+    authorize @item.becomes(Item)
     redirect_to item_path(@item) << '#edit'
   end
   
   def update
     @item = get_by_uid params[:uid]
+    authorize @item
     @item.update item_params
     redirect_to item_path(@item)
   end
   
   def new
     @item = Item.new
+    authorize @item.becomes(Item)
   end
   
   def create
     @item = Item.new(item_params)
-    
+    authorize @item.becomes(Item)
     if @item.save
       # Use item_path instead of plain @item to force use of item controller
       # instead of type-specific controller (not implemented (yet))
@@ -41,6 +46,7 @@ class ItemsController < ApplicationController
   end
   
   def destroy
+    authorize params[:uid]
     Item.destroy_all(uid: params[:uid])
     redirect_to items_path
   end
